Added mod commands logging into dedicated channel
